requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic>=2.8,<2.9
sqlalchemy>=2.0,<2.1
pymysql>=1.1,<1.2
cryptography>=41.0,<42.0
python-dotenv>=1.0,<1.1
